requirements.txt
streamlit
requests
python-dotenv
streamlit
openai
faiss-cpu
numpy